Extract PathConditionRoad to be usable by other classes
